feat: access token 받는 방식 변경(#146)
