+++ src/services/authApi.js
@@ -1,31 +1,4 @@
-import axios from "axios";
-// import { fetchGetUserData } from "./userApi";
-
-export const api = axios.create({
-  baseURL: import.meta.env?.VITE_API_BASE_URL,
-  withCredentials: true, // Refresh 쿠키 전송/수신 -> 서버 CORS 정책 : Access-Control-Allow-Credentials: true
-  headers: {
-    "Content-Type": "application/json",
-  },
-});
-
-api.interceptors.request.use(
-  (config) => {
-    return config;
-  },
-  (err) => {
-    Promise.reject(err);
-  }
-);
-
-api.interceptors.response.use(
-  (res) => res,
-  async (error) => {
-    if (error.response?.status === 401) {
-      await api.post("/api/token/refresh");
-    }
-  }
-);
+import { api } from "./api";
 
 export const fetchSignin = async (email, password, name) => {
   const userData = {
